backend/requirements.txt
Django==2.2.16
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework==3.12.4
djoser==2.1.0
django-filter==21.1
django-colorfield==0.7.2
drf-extra-fields==3.4.0
psycopg2-binary==2.9.3